+++ CHANGELOG.md
@@ -56,6 +56,8 @@
 - Rename `CosmCache` to just `cosmwasm_vm::Cache` and add `CacheOptions` to
   configure it.
 - Rename `Extern` to `Backend`.
+- Rename `mock_dependencies` to `mock_backend` and
+  `mock_dependencies_with_balances` to `mock_backend_with_balances`.
 
 ## 0.11.2 (2020-10-26)